requirements.txt
# Core Framework
streamlit>=1.39.0
langchain>=0.3.0
langchain-core>=0.3.0

# LLM & Embeddings
langchain-openai>=0.2.0
openai>=1.50.0

# Vector Database
langchain-chroma>=0.1.4
chromadb>=0.5.0

# Web Search Integration
langchain-tavily>=0.2.0
tavily-python>=0.5.0

# Document Processing
pypdf>=4.3.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-multipart>=0.0.6

# Text Processing & Embeddings
sentence-transformers>=3.0.0
rank-bm25>=0.2.2

# Vector Search & Retrieval
faiss-cpu>=1.8.0
numpy>=1.24.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Utilities
requests>=2.31.0
typing-extensions>=4.8.0
pandas>=2.1.0

# UI Enhancements
streamlit-chat>=0.1.1
plotly>=5.17.0 
dotenv>=0.21.0